requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.0
configparser==3.5.0
coreapi==2.1.1
decorator==4.0.10
dj-database-url==0.3.0
Django==1.8.5
django-celery==3.1.17
django-crispy-forms==1.6.1
django-filter==1.0.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
inflection==0.3.1
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
jsonschema==2.5.1
kombu==3.0.37
Markdown==2.6.3
marshmallow==2.10.4
mysqlclient==1.3.9
openapi==0.5.0
openapi-codec==1.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.0.2
pytz==2016.10
requests==2.12.4
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
tornado==4.4.2
traitlets==4.3.1
uritemplate==3.0.0
vine==1.1.3
wcwidth==0.1.7
